Создан PageTemplate для страниц пользователя и форума, внесены правки по замечаниям ревью
