+++ packages/client/src/pages/profile/components/ChangeDataForm/index.ts
@@ -1,0 +1,38 @@
+export const fields = {
+  login: {
+    label: 'Логин',
+    placeholder: 'Логин',
+    type: 'text',
+    autocomplete: 'username',
+  },
+  email: {
+    label: 'Почта',
+    placeholder: 'Почта',
+    type: 'email',
+    autocomplete: 'email',
+  },
+  first_name: {
+    label: 'Имя',
+    placeholder: 'Имя',
+    type: 'text',
+    autocomplete: 'given-name',
+  },
+  second_name: {
+    label: 'Фамилия',
+    placeholder: 'Фамилия',
+    type: 'text',
+    autocomplete: 'family-name',
+  },
+  display_name: {
+    label: 'Имя в чате',
+    placeholder: 'Имя в чате',
+    type: 'text',
+    autocomplete: 'nickname',
+  },
+  phone: {
+    label: 'Телефон',
+    placeholder: '+7 (000) 000-00-00',
+    type: 'tel',
+    autocomplete: 'tel',
+  },
+}
