+++ packages/client/src/pages/profile/components/ChangePasswordForm/index.ts
@@ -1,0 +1,22 @@
+export const fields = {
+  oldPassword: {
+    label: 'Старый пароль',
+    placeholder: 'Старый пароль',
+    type: 'password',
+    value: '',
+    autocomplete: 'current-password',
+  },
+  newPassword: {
+    label: 'Новый пароль',
+    placeholder: 'Новый пароль',
+    type: 'password',
+    autocomplete: 'new-password',
+  },
+  newPasswordAgain: {
+    label: 'Новый пароль еще раз',
+    placeholder: 'Новый пароль еще раз',
+    type: 'password',
+    value: '',
+    autocomplete: 'new-password',
+  },
+}
