+++ packages/client/src/pages/profile/styled.ts
@@ -1,21 +1,4 @@
 import styled from '@emotion/styled'
-import { Container } from '@mui/material'
-
-type StyledContainerProps = {
-  maxWidth: string
-  component: string
-}
-
-export const StyledContainer = styled(Container)<StyledContainerProps>(() => ({
-  display: 'flex',
-  justifyContent: 'center',
-  minHeight: '700px',
-  backgroundColor: '#fff',
-  border: '1px solid #ddd',
-  padding: '72px',
-  borderRadius: '8px',
-  alignItems: 'center',
-}))
 
 export const StyledForm = styled.form`
   margin: 20px auto;
@@ -27,10 +10,4 @@ export const ButtonsContainer = styled.div`
   flex-direction: column;
   gap: 0.5rem;
   margin: 40px auto;
-`
-
-export const StyledWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
 `
